requirements.txt
# Pipeline de datos
requests>=2.31.0
pyalex>=0.13
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
python-Levenshtein>=0.21.0

# Modelo
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
accelerate>=0.25.0
evaluate>=0.4.0

# Métricas
jiwer>=3.0.0

# API (opcional)
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0

# Utilidades
python-dotenv>=1.0.0
PyYAML>=6.0.1
tqdm>=4.66.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
